requirements.txt
python-ndn==0.5.0
